+++ backend/src/main/java/com/ssafy/trip/repository/heart/HeartRepository.java
@@ -1,4 +1,4 @@
-package com.ssafy.trip.repository;
+package com.ssafy.trip.repository.heart;
 
 import com.ssafy.trip.Entity.Attraction;
 import com.ssafy.trip.Entity.Heart;
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface HeartRepository extends JpaRepository<Heart,Integer> {
+public interface HeartRepository extends JpaRepository<Heart,Integer> , HeartRepositoryCustom {
 
       Optional<Heart> findByMemberAndAttraction(Member member , Attraction attraction);
