+++ backend/src/main/java/com/ssafy/trip/repository/heart/HeartRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.ssafy.trip.repository.heart;
+
+import java.util.List;
+
+public interface HeartRepositoryCustom {
+
+    List<Integer> findBySearch(String userId);
+
+}
